pyproject.toml
[build-system]
requires = [
    "setuptools>=69.0",
    "wheel"
]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["src"]
include = ["melo_benchmark*"]
exclude = ["tests*"]

[tool.setuptools.package-data]
skills_core = [
    "resources/*.csv",
]

[project]
name = "melo-benchmark"
version = "1.0.0"
description = "Multilingual Entity Linking for Occupations (MELO) Benchmark."
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "matplotlib>=3.9.2",
    "numpy>=1.21.1",
    "openai>=1.30.3",
    "pandas>=2.2.2",
    "psutil>=5.9.8",
    "python-dotenv>=1.0.1",
    "rank-bm25>=0.2.2",
    "rapidfuzz>=3.9.4",
    "scikit-learn==1.4.0",
    "scipy>=1.14.1",
    "seaborn>=0.13.2",
    "spacy>=3.7.5"
]
classifiers = [
    "Programming Language :: Python :: 3",
    "Operating System :: OS Independent",
]

[project.urls]
Homepage = "https://github.com/federetyk/melo-benchmark"
Documentation = "https://github.com/federetyk/melo-benchmark/docs"
Repository = "https://github.com/federetyk/melo-benchmark"
Issues = "https://github.com/federetyk/melo-benchmark/issues"
